# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/class-groups": {
      "get": {
        "operationId": "ClassGroupsController_findAll",
        "parameters": [
          {
            "name": "schoolYearId",
            "required": false,
            "in": "query",
            "description": "Filter by school year id",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "gradeLevel",
            "required": false,
            "in": "query",
            "description": "Filter by grade level (1-11)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "section",
            "required": false,
            "in": "query",
            "description": "Filter by two-digit section code",
            "schema": {
              "example": "01",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Full-text search term",
            "schema": {
              "example": "algebra",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1+)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size (1-100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "class-groups"
        ]
      },
      "post": {
        "operationId": "ClassGroupsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassGroupDto"
              },
              "examples": {
                "default": {
                  "summary": "Create class group",
                  "value": {
                    "schoolYearId": 1,
                    "gradeLevel": 10,
                    "section": "01",
                    "defaultClassroomId": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "class-groups"
        ]
      }
    },
    "/class-groups/{id}": {
      "get": {
        "operationId": "ClassGroupsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "class-groups"
        ]
      },
      "patch": {
        "operationId": "ClassGroupsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassGroupDto"
              },
              "examples": {
                "default": {
                  "summary": "Update class group",
                  "value": {
                    "gradeLevel": 11,
                    "section": "02"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "class-groups"
        ]
      },
      "delete": {
        "operationId": "ClassGroupsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "class-groups"
        ]
      }
    },
    "/classrooms": {
      "get": {
        "operationId": "ClassroomsController_findAll",
        "parameters": [
          {
            "name": "building",
            "required": false,
            "in": "query",
            "description": "Filter by building name",
            "schema": {
              "example": "Main Building",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Keyword search within name",
            "schema": {
              "example": "Room",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1+)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size (1-100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "classrooms"
        ]
      },
      "post": {
        "operationId": "ClassroomsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassroomDto"
              },
              "examples": {
                "default": {
                  "summary": "Create classroom",
                  "value": {
                    "name": "Room 101",
                    "building": "Main Building",
                    "capacity": 30
                  }
                }
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "classrooms"
        ]
      }
    },
    "/classrooms/{id}": {
      "get": {
        "operationId": "ClassroomsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "classrooms"
        ]
      },
      "patch": {
        "operationId": "ClassroomsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassroomDto"
              },
              "examples": {
                "default": {
                  "summary": "Update classroom capacity",
                  "value": {
                    "capacity": 28
                  }
                }
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "classrooms"
        ]
      },
      "delete": {
        "operationId": "ClassroomsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "classrooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "School Man API",
    "description": "Generated OpenAPI specification for School Man backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateClassGroupDto": {
        "type": "object",
        "properties": {
          "schoolYearId": {
            "type": "number",
            "example": 1,
            "description": "Existing school year identifier"
          },
          "gradeLevel": {
            "type": "number",
            "example": 10,
            "description": "Grade level (1-11)"
          },
          "section": {
            "type": "string",
            "example": "01",
            "description": "Two-digit section code"
          },
          "defaultClassroomId": {
            "type": "number",
            "example": 3,
            "description": "Optional default classroom id"
          }
        },
        "required": [
          "schoolYearId",
          "gradeLevel",
          "section"
        ]
      },
      "UpdateClassGroupDto": {
        "type": "object",
        "properties": {}
      },
      "CreateClassroomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Room 101",
            "description": "Unique classroom name"
          },
          "building": {
            "type": "string",
            "example": "Main Building",
            "description": "Building where the classroom is located"
          },
          "capacity": {
            "type": "number",
            "example": 30,
            "description": "Maximum capacity of the classroom"
          }
        },
        "required": [
          "name",
          "capacity"
        ]
      },
      "UpdateClassroomDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}